Extract WorkSequence into separate files
