+++ src/work/WorkSequence.cpp
@@ -1,5 +1,95 @@
-//
-// Created by Marta Lokhava on 1/25/19.
-//
+// Copyright 2019 Stellar Development Foundation and contributors. Licensed
+// under the Apache License, Version 2.0. See the COPYING file at the root
+// of this distribution or at http://www.apache.org/licenses/LICENSE-2.0
 
 #include "WorkSequence.h"
+
+namespace stellar
+{
+
+WorkSequence::WorkSequence(Application& app, std::string name,
+                           std::vector<std::shared_ptr<BasicWork>> sequence,
+                           size_t maxRetries)
+    : BasicWork(app, std::move(name), maxRetries)
+    , mSequenceOfWork(sequence)
+    , mNextInSequence(mSequenceOfWork.begin())
+{
+}
+
+BasicWork::State
+WorkSequence::onRun()
+{
+    if (mNextInSequence == mSequenceOfWork.end())
+    {
+        // Completed all the work
+        return State::WORK_SUCCESS;
+    }
+
+    auto w = *mNextInSequence;
+    assert(w);
+    if (!mCurrentExecuting)
+    {
+        w->startWork(wakeSelfUpCallback());
+        mCurrentExecuting = w;
+    }
+    else
+    {
+        auto state = w->getState();
+        if (state == State::WORK_SUCCESS)
+        {
+            mCurrentExecuting = nullptr;
+            ++mNextInSequence;
+            return this->onRun();
+        }
+        else if (state != State::WORK_RUNNING)
+        {
+            return state;
+        }
+    }
+    w->crankWork();
+    return State::WORK_RUNNING;
+}
+
+bool
+WorkSequence::onAbort()
+{
+    if (mCurrentExecuting && !mCurrentExecuting->isDone())
+    {
+        // Wait for the current work in sequence to finish aborting
+        mCurrentExecuting->crankWork();
+        return false;
+    }
+    return true;
+}
+
+void
+WorkSequence::onReset()
+{
+    assert(std::all_of(
+        mSequenceOfWork.cbegin(), mNextInSequence,
+        [](std::shared_ptr<BasicWork> const& w) { return w->isDone(); }));
+    mNextInSequence = mSequenceOfWork.begin();
+    mCurrentExecuting.reset();
+}
+
+std::string
+WorkSequence::getStatus() const
+{
+    if (!isDone() && mNextInSequence != mSequenceOfWork.end())
+    {
+        return (*mNextInSequence)->getStatus();
+    }
+    return BasicWork::getStatus();
+}
+
+void
+WorkSequence::shutdown()
+{
+    if (mCurrentExecuting)
+    {
+        mCurrentExecuting->shutdown();
+    }
+
+    BasicWork::shutdown();
+}
+}
